test: 인증 관련 테스트와 로직(삭제) 로직 테스트 분리
인증관련 테스트와 로직 관련 테스트 분리
