+++ src/test/java/gift/controller/auth/AuthControllerTest.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.jayway.jsonpath.JsonPath;
 import gift.dto.LoginRequest;
 import gift.dto.RegisterRequest;
+import gift.service.MemberService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,7 +13,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -25,6 +25,10 @@ class AuthControllerTest {
     private MockMvc mockMvc;
     @Autowired
     private ObjectMapper objectMapper;
+    @Autowired
+    private MemberService memberService;
+    @Autowired
+    private AuthInterceptor authInterceptor;
 
     @Test
     @DisplayName("빈 이름으로 회원가입 요청하기")
@@ -110,8 +114,6 @@ class AuthControllerTest {
                 .contentType(MediaType.APPLICATION_JSON)
                 .content(objectMapper.writeValueAsString(new RegisterRequest("테스트", "test@naver.com", "testPassword", "MEMBER"))));
 
-        register.andExpect(status().isOk());
-
         var login = mockMvc.perform(post("/api/members/login")
                 .contentType(MediaType.APPLICATION_JSON)
                 .content(objectMapper.writeValueAsString(new LoginRequest("test@naver.com", "testPasswordWrong"))));
@@ -120,23 +122,17 @@ class AuthControllerTest {
 
         var result = register.andExpect(status().isOk()).andReturn();
         var responseContent = result.getResponse().getContentAsString();
-        var token = JsonPath.parse(responseContent).read("$.token");
+        var token = JsonPath.parse(responseContent).read("$.token").toString();
 
-        var deleted = mockMvc.perform(delete("/api/members")
-                .contentType(MediaType.APPLICATION_JSON)
-                .header("Authorization", "Bearer " + token));
-
-        deleted.andExpect(status().isNoContent());
+        memberService.deleteMember(authInterceptor.getMemberIdWithToken(token));
     }
 
     @Test
     @DisplayName("정상적으로 회원가입 후 로그인 요청하기")
     void registerAndLoginSuccess() throws Exception {
-        var register = mockMvc.perform(post("/api/members/register")
+        mockMvc.perform(post("/api/members/register")
                 .contentType(MediaType.APPLICATION_JSON)
                 .content(objectMapper.writeValueAsString(new RegisterRequest("테스트", "test@naver.com", "testPassword", "MEMBER"))));
-
-        register.andExpect(status().isOk());
 
         var login = mockMvc.perform(post("/api/members/login")
                 .contentType(MediaType.APPLICATION_JSON)
@@ -144,12 +140,8 @@ class AuthControllerTest {
 
         var result = login.andExpect(status().isOk()).andReturn();
         var responseContent = result.getResponse().getContentAsString();
-        var token = JsonPath.parse(responseContent).read("$.token");
+        var token = JsonPath.parse(responseContent).read("$.token").toString();
 
-        var deleted = mockMvc.perform(delete("/api/members")
-                .contentType(MediaType.APPLICATION_JSON)
-                .header("Authorization", "Bearer " + token));
-
-        deleted.andExpect(status().isNoContent());
+        memberService.deleteMember(authInterceptor.getMemberIdWithToken(token));
     }
 }
