+++ src/test/java/gift/service/auth/AuthServiceTest.java
@@ -5,10 +5,8 @@ import gift.dto.LoginRequest;
 import gift.dto.RegisterRequest;
 import gift.exception.DuplicatedEmailException;
 import gift.exception.InvalidLoginInfoException;
-import gift.exception.NotFoundElementException;
 import gift.model.MemberRole;
 import gift.service.MemberService;
-import gift.service.auth.AuthService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -77,21 +75,5 @@ class AuthServiceTest {
 
         var id = authInterceptor.getMemberIdWithToken(auth.token());
         memberService.deleteMember(id);
-    }
-
-    @Test
-    @DisplayName("회원 탈퇴하기 - 성공")
-    void deleteMemberSuccess() {
-        var registerRequest = new RegisterRequest("테스트", "test@naver.com", "testPassword", "MEMBER");
-        var loginRequest = new LoginRequest("test@naver.com", "testPassword");
-        var auth = authService.register(registerRequest);
-        var id = authInterceptor.getMemberIdWithToken(auth.token());
-        var loginAuth = authService.login(loginRequest);
-
-        Assertions.assertThat(auth.token()).isEqualTo(loginAuth.token());
-
-        memberService.deleteMember(id);
-
-        Assertions.assertThatThrownBy(() -> authService.login(loginRequest)).isInstanceOf(NotFoundElementException.class);
     }
 }
